Revised HTML5 Basic starter theme
Joachim Kudish was Kind enough to review my theme and so a few changes
were made.
Reset.css was removed and replaced by normalize.css
HTML5 js was moved to it's own folder
The header.php was modified by getting rid of all the seo stuff - that
can be done with a plugin
The functions.php file was cleaned up and optimized.
I also ran the theme-check plugin and fixed a few niggly things